@@ -3,18 +3,24 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@emotion/react": "^11.9.0",
+    "@emotion/styled": "^11.8.1",
     "@material-ui/core": "^4.11.0",
     "@material-ui/icons": "^4.9.1",
+    "@mui/icons-material": "^5.6.2",
+    "@mui/material": "^5.6.3",
+    "@mui/styled-engine-sc": "^5.6.1",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
     "clsx": "^1.1.1",
     "clx": "^1.0.0",
-    "firebase": "^7.20.0",
-    "react": "^16.13.1",
+    "firebase": "^9.6.11",
+    "react": "^18.1.0",
     "react-dom": "^16.13.1",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "3.4.3"
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "3.4.3",
+    "styled-components": "^5.3.5"
   },
   "scripts": {
     "start": "react-scripts start",
